move DBConnection file util to db
