+++ src/main/java/chess/db/DBConnection.java
@@ -1,4 +1,4 @@
-package chess.util;
+package chess.db;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
